+++ src/util.c
@@ -31,7 +31,7 @@
 
 static const char html_header[] = "HTTP/1.1 200 OK\r\nContent-Type: text/plain\r\n";
 
-int ambitv_util_append_ptr_to_list(void*** list_ptr, int idx, long* len_ptr, void* ptr)
+int ambitv_util_append_ptr_to_list(void*** list_ptr, int idx, intptr_t* len_ptr, void* ptr)
 {
 	if (NULL == *list_ptr)
 	{
@@ -49,11 +49,11 @@ int ambitv_util_append_ptr_to_list(void*** list_ptr, int idx, long* len_ptr, voi
 	return ++idx;
 }
 
-int ambitv_parse_led_string(const char* str, long** out_ptr, long* out_len)
+int ambitv_parse_led_string(const char* str, intptr_t** out_ptr, intptr_t* out_len)
 {
 	int a, b, c, idx = 0, skip = 0, *s = &a;
-	long ilen = 0;
-	long* list = NULL;
+	intptr_t ilen = 0;
+	intptr_t* list = NULL;
 
 	a = b = 0;
 	while (1)
